format fixes

+++ src/access_log.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: access_log.cc,v 1.71 2001/10/17 12:41:49 hno Exp $
+ * $Id: access_log.cc,v 1.72 2001/10/24 06:55:43 hno Exp $
  *
  * DEBUG: section 46    Access Log
  * AUTHOR: Duane Wessels
@@ -242,14 +242,14 @@ accessLogSquid(AccessLogEntry * al)
 	client = inet_ntoa(al->cache.caddr);
     user = accessLogFormatName(al->cache.authuser ?
 	al->cache.authuser : al->cache.rfc931);
-    logfilePrintf(logfile, "%9d.%03d %6d %s %s/%03d %d %s %s %s %s%s/%s %s",
+    logfilePrintf(logfile, "%9d.%03d %6d %s %s/%03d %ld %s %s %s %s%s/%s %s",
 	(int) current_time.tv_sec,
 	(int) current_time.tv_usec / 1000,
 	al->cache.msec,
 	client,
 	log_tags[al->cache.code],
 	al->http.code,
-	al->cache.size,
+	(long int) al->cache.size,
 	al->private.method_str,
 	al->url,
 	user ? user : dash_str,
@@ -271,7 +271,7 @@ accessLogCommon(AccessLogEntry * al)
 	client = inet_ntoa(al->cache.caddr);
     user1 = accessLogFormatName(al->cache.authuser);
     user2 = accessLogFormatName(al->cache.rfc931);
-    logfilePrintf(logfile, "%s %s %s [%s] \"%s %s HTTP/%d.%d\" %d %d %s:%s",
+    logfilePrintf(logfile, "%s %s %s [%s] \"%s %s HTTP/%d.%d\" %d %ld %s:%s",
 	client,
 	user2 ? user2 : dash_str,
 	user1 ? user1 : dash_str,
@@ -280,7 +280,7 @@ accessLogCommon(AccessLogEntry * al)
 	al->url,
 	al->http.version.major, al->http.version.minor,
 	al->http.code,
-	al->cache.size,
+	(long int) al->cache.size,
 	log_tags[al->cache.code],
 	hier_strings[al->hier.code]);
     safe_free(user1);
